+++ dune/stuff/common/fixed_map.hh
@@ -148,7 +148,7 @@ public:
   {
   }
 
-  int getIdx(const key_type& key) const
+  long getIdx(const key_type& key) const
   {
     const auto it = std::find_if(map_.begin(), map_.end(), [&](const value_type& val) { return val.first == key; });
     if (it == map_.end())
